worker/requirements.txt
faststream==0.5.33
faststream[cli]
aio_pika==9.5.4
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2==2.9.10
asyncpg==0.30.0
python-dotenv==1.0.1
